backend/requirements.txt
# Requirements for the backend application

# Web server
fastapi

# LLM server
llama-cpp-python[all]

# OpenAI Agent SDK
openai-agents

# Environment variable management
python-dotenv

# Pydantic for data validation
pydantic

# Some tools
gson